split probin.f90 into probin.f90 and probin_gmres.f90.  Split the namelist into probin and probin_gmres.
NOTE: in order for the code to work I still need to modify the inputs files so the code knows there are multiple namelists.  I may wait on doing this until that "what goes where" is settled a bit more.  in the meantime I will check in (shortly) and maintain a sample working inputs file